No need to set IPV6_CHECKSUM for an ICMPv6 socket
According to RFC 3542, the kernel will calculate and insert the ICMPv6
checksum for ICMPv6 raw sockets, since this checksum is mandatory, so
setting IPV6_CHECKSUM explictly is useless and will fail.

To be on the safe side, we follow the example of iputils and set
IPV6_CHECKSUM and simply lgo the failure.

Signed-off-by: Dimitri Papadopoulos <3350651-DimitriPapadopoulos@users.noreply.gitlab.com>