Make path to ocserv-fw script conform to config prefix
The path to ocserv-fw script was hard-coded and didn't
respect the ${exec_prefix} variable in the build process.
This caused the script to be installed in a different
location than the hard-coded default path so ocserv would
error if it tried to execute it. Now the path is set using
a build flag so it can be rooted at the ${exec_prefix}
with the rest of the application. An if guard has been
added to the original macro line as a fallback and to assist
build and editing tools.

Signed-off-by: Lee Keitel <lee@keitel.xyz>